uint32_t for microseconds of subjective billing only allows for 1.2hrs before wrapping. Use int64_t instead which matches type used for objective billing.

+++ libraries/chain/include/eosio/chain/controller.hpp
@@ -152,7 +152,7 @@ namespace eosio { namespace chain {
          transaction_trace_ptr push_transaction( const transaction_metadata_ptr& trx,
                                                  fc::time_point deadline, fc::microseconds max_transaction_time,
                                                  uint32_t billed_cpu_time_us, bool explicit_billed_cpu_time,
-                                                 uint32_t subjective_cpu_bill_us );
+                                                 int64_t subjective_cpu_bill_us );
 
          /**
           * Attempt to execute a specific transaction in our deferred trx database
